Move PostBuildSteps to its own .java file
